requirements.txt
numpy
scikit-learn
pandas
kaggle
git+https://github.com/gbolmier/funk-svd
scikit-surprise
pytorch
lightning
